Add Branch

+++ AL-Game/data/scripts/system/handlers/ai/classNpc/DrakanPriestAI2.java
@@ -14,7 +14,10 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package ai.classNpc;
+
+import ai.AggressiveNpcAI2;
 
 import com.aionemu.commons.utils.Rnd;
 import com.aionemu.gameserver.ai2.AIName;
@@ -26,10 +29,9 @@ import com.aionemu.gameserver.spawnengine.SpawnEngine;
 import com.aionemu.gameserver.spawnengine.VisibleObjectSpawner;
 import com.aionemu.gameserver.world.knownlist.Visitor;
 
-import ai.AggressiveNpcAI2;
-
 /**
  * @author Cheatkiller
+ *
  */
 @AIName("xdrakanpriest")
 public class DrakanPriestAI2 extends AggressiveNpcAI2 {
@@ -64,7 +66,6 @@ public class DrakanPriestAI2 extends AggressiveNpcAI2 {
 
 	private void despawnServant() {
 		getOwner().getKnownList().doOnAllNpcs(new Visitor<Npc>() {
-
 			@Override
 			public void visit(Npc object) {
 				Npc healServant = getPosition().getWorldMapInstance().getNpc(282988);
@@ -78,7 +79,7 @@ public class DrakanPriestAI2 extends AggressiveNpcAI2 {
 	private void rndSpawn(int npcId, int count) {
 		for (int i = 0; i < count; i++) {
 			SpawnTemplate template = rndSpawnInRange(npcId);
-			VisibleObjectSpawner.spawnEnemyServant(template, getOwner().getInstanceId(), getOwner(), getOwner().getLevel());
+			VisibleObjectSpawner.spawnEnemyServant(template, getOwner().getInstanceId(), getOwner(), (byte) getOwner().getLevel());
 		}
 	}
 
@@ -86,6 +87,7 @@ public class DrakanPriestAI2 extends AggressiveNpcAI2 {
 		float direction = Rnd.get(0, 199) / 100f;
 		float x1 = (float) (Math.cos(Math.PI * direction) * 5);
 		float y1 = (float) (Math.sin(Math.PI * direction) * 5);
-		return SpawnEngine.addNewSingleTimeSpawn(getPosition().getMapId(), npcId, getPosition().getX() + x1, getPosition().getY() + y1, getPosition().getZ(), getPosition().getHeading());
+		return SpawnEngine.addNewSingleTimeSpawn(getPosition().getMapId(), npcId, getPosition().getX() + x1, getPosition().getY() + y1, getPosition().getZ(),
+				getPosition().getHeading());
 	}
 }
