+++ AL-Game/src/com/aionemu/gameserver/geoEngine/scene/Node.java
@@ -14,13 +14,8 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.aionemu.gameserver.geoEngine.scene;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.logging.Logger;
-
-import javax.activation.UnsupportedDataTypeException;
 
 import com.aionemu.gameserver.geoEngine.bounding.BoundingVolume;
 import com.aionemu.gameserver.geoEngine.collision.Collidable;
@@ -30,9 +25,16 @@ import com.aionemu.gameserver.geoEngine.math.Matrix3f;
 import com.aionemu.gameserver.geoEngine.math.Ray;
 import com.aionemu.gameserver.geoEngine.math.Vector3f;
 
+import javax.activation.UnsupportedDataTypeException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.logging.Logger;
+
 /**
- * <code>Node</code> defines an internal node of a scene graph. The internal node maintains a collection of children and handles merging said children into a single bound to allow for very fast
- * culling of multiple nodes. Node allows for any number of children to be attached.
+ * <code>Node</code> defines an internal node of a scene graph. The internal
+ * node maintains a collection of children and handles merging said children
+ * into a single bound to allow for very fast culling of multiple nodes. Node
+ * allows for any number of children to be attached.
  *
  * @author Mark Powell
  * @author Gregg Patton
@@ -54,18 +56,21 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * Constructor instantiates a new <code>Node</code> with a default empty list for containing children.
+	 * Constructor instantiates a new <code>Node</code> with a default empty
+	 * list for containing children.
 	 *
 	 * @param name
-	 *            the name of the scene element. This is required for identification and comparision purposes.
+	 *            the name of the scene element. This is required for
+	 *            identification and comparision purposes.
 	 */
 	public Node(String name) {
 		super(name);
-		collisionFlags = CollisionIntention.ALL.getId();
+		collisionFlags = (short) CollisionIntention.ALL.getId();
 	}
 
 	/**
-	 * <code>getQuantity</code> returns the number of children this node maintains.
+	 * <code>getQuantity</code> returns the number of children this node
+	 * maintains.
 	 *
 	 * @return the number of children this node maintains.
 	 */
@@ -74,7 +79,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>getTriangleCount</code> returns the number of triangles contained in all sub-branches of this node that contain geometry.
+	 * <code>getTriangleCount</code> returns the number of triangles contained
+	 * in all sub-branches of this node that contain geometry.
 	 *
 	 * @return the triangle count of this branch.
 	 */
@@ -91,7 +97,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>getVertexCount</code> returns the number of vertices contained in all sub-branches of this node that contain geometry.
+	 * <code>getVertexCount</code> returns the number of vertices contained in
+	 * all sub-branches of this node that contain geometry.
 	 *
 	 * @return the vertex count of this branch.
 	 */
@@ -108,7 +115,9 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>attachChild</code> attaches a child to this node. This node becomes the child's parent. The current number of children maintained is returned. <br>
+	 * <code>attachChild</code> attaches a child to this node. This node becomes
+	 * the child's parent. The current number of children maintained is
+	 * returned. <br>
 	 * If the child already had a parent it is detached from that former parent.
 	 *
 	 * @param child
@@ -134,7 +143,9 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>attachChildAt</code> attaches a child to this node at an index. This node becomes the child's parent. The current number of children maintained is returned. <br>
+	 * <code>attachChildAt</code> attaches a child to this node at an index.
+	 * This node becomes the child's parent. The current number of children
+	 * maintained is returned. <br>
 	 * If the child already had a parent it is detached from that former parent.
 	 *
 	 * @param child
@@ -160,7 +171,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>detachChild</code> removes a given child from the node's list. This child will no longe be maintained.
+	 * <code>detachChild</code> removes a given child from the node's list. This
+	 * child will no longe be maintained.
 	 *
 	 * @param child
 	 *            the child to remove.
@@ -183,7 +195,9 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>detachChild</code> removes a given child from the node's list. This child will no longe be maintained. Only the first child with a matching name is removed.
+	 * <code>detachChild</code> removes a given child from the node's list. This
+	 * child will no longe be maintained. Only the first child with a matching
+	 * name is removed.
 	 *
 	 * @param childName
 	 *            the child to remove.
@@ -205,7 +219,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>detachChildAt</code> removes a child at a given index. That child is returned for saving purposes.
+	 * <code>detachChildAt</code> removes a child at a given index. That child
+	 * is returned for saving purposes.
 	 *
 	 * @param index
 	 *            the index of the child to be removed.
@@ -220,7 +235,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>detachAllChildren</code> removes all children attached to this node.
+	 * <code>detachAllChildren</code> removes all children attached to this
+	 * node.
 	 */
 	public void detachAllChildren() {
 		for (int i = children.size() - 1; i >= 0; i--) {
@@ -259,7 +275,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * <code>getChild</code> returns the first child found with exactly the given name (case sensitive.)
+	 * <code>getChild</code> returns the first child found with exactly the
+	 * given name (case sensitive.)
 	 *
 	 * @param name
 	 *            the name of the child to retrieve. If null, we'll return null.
@@ -274,8 +291,7 @@ public class Node extends Spatial implements Cloneable {
 			Spatial child = children.get(x);
 			if (name.equals(child.getName())) {
 				return child;
-			}
-			else if (child instanceof Node) {
+			} else if (child instanceof Node) {
 				Spatial out = ((Node) child).getChild(name);
 				if (out != null) {
 					return out;
@@ -286,7 +302,8 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * determines if the provided Spatial is contained in the children list of this node.
+	 * determines if the provided Spatial is contained in the children list of
+	 * this node.
 	 *
 	 * @param spat
 	 *            the child object to look for.
@@ -340,7 +357,8 @@ public class Node extends Spatial implements Cloneable {
 			Spatial child = children.get(i);
 			if (child instanceof Geometry) {
 
-				// not used materialIds do not have collision intention for materials set
+				// not used materialIds do not have collision intention for
+				// materials set
 				// not all material meshes have physical collisions set
 				// TODO: implement event mesh collisions
 				if ((child.getIntentions() & results.getIntentions()) == 0 || (child.getIntentions() & CollisionIntention.EVENT.getId()) != 0) {
@@ -359,11 +377,13 @@ public class Node extends Spatial implements Cloneable {
 	}
 
 	/**
-	 * Returns flat list of Spatials implementing the specified class AND with name matching the specified pattern.
-	 * </P>
+	 * Returns flat list of Spatials implementing the specified class AND with
+	 * name matching the specified pattern. </P>
 	 * <p/>
-	 * Note that we are <i>matching</i> the pattern, therefore the pattern must match the entire pattern (i.e. it behaves as if it is sandwiched between "^" and "$"). You can set regex modes, like
-	 * case insensitivity, by using the (?X) or (?X:Y) constructs.
+	 * Note that we are <i>matching</i> the pattern, therefore the pattern must
+	 * match the entire pattern (i.e. it behaves as if it is sandwiched between
+	 * "^" and "$"). You can set regex modes, like case insensitivity, by using
+	 * the (?X) or (?X:Y) constructs.
 	 * </P>
 	 * <p/>
 	 * By design, it is always safe to code loops like: <CODE><PRE>
@@ -371,15 +391,20 @@ public class Node extends Spatial implements Cloneable {
 	 * </PRE></CODE>
 	 * </P>
 	 * <p/>
-	 * "Descendants" does not include self, per the definition of the word. To test for descendants AND self, you must do a <code>node.matches(aClass, aRegex)</code> +
+	 * "Descendants" does not include self, per the definition of the word. To
+	 * test for descendants AND self, you must do a
+	 * <code>node.matches(aClass, aRegex)</code> +
 	 * <code>node.descendantMatches(aClass, aRegex)</code>.
 	 * <p/>
 	 *
 	 * @param spatialSubclass
-	 *            Subclass which matching Spatials must implement. Null causes all Spatials to qualify.
+	 *            Subclass which matching Spatials must implement. Null causes
+	 *            all Spatials to qualify.
 	 * @param nameRegex
-	 *            Regular expression to match Spatial name against. Null causes all Names to qualify.
-	 * @return Non-null, but possibly 0-element, list of matching Spatials (also Instances extending Spatials).
+	 *            Regular expression to match Spatial name against. Null causes
+	 *            all Names to qualify.
+	 * @return Non-null, but possibly 0-element, list of matching Spatials (also
+	 *         Instances extending Spatials).
 	 * @see java.util.regex.Pattern
 	 * @see Spatial#matches(Class<? extends Spatial>, String)
 	 */
@@ -438,8 +463,7 @@ public class Node extends Spatial implements Cloneable {
 				if (resultBound != null) {
 					// merge current world bound with child world bound
 					resultBound.mergeLocal(child.getWorldBound());
-				}
-				else {
+				} else {
 					// set world bound to first non-null child world bound
 					if (child.getWorldBound() != null) {
 						resultBound = child.getWorldBound().clone(this.worldBound);
@@ -452,7 +476,10 @@ public class Node extends Spatial implements Cloneable {
 
 	/*
 	 * (non-Javadoc)
-	 * @see aionjHungary.geoEngine.scene.Spatial#setTransform(aionjHungary.geoEngine.math.Matrix3f, aionjHungary.geoEngine.math.Vector3f)
+	 *
+	 * @see
+	 * aionjHungary.geoEngine.scene.Spatial#setTransform(aionjHungary.geoEngine
+	 * .math.Matrix3f, aionjHungary.geoEngine.math.Vector3f)
 	 */
 	@Override
 	public void setTransform(Matrix3f rotation, Vector3f loc, float scale) {
@@ -471,11 +498,9 @@ public class Node extends Spatial implements Cloneable {
 			if (spatial instanceof Geometry) {
 				Geometry geom = new Geometry(spatial.getName(), ((Geometry) spatial).getMesh());
 				node.attachChild(geom);
-			}
-			else if (spatial instanceof Node) {
+			} else if (spatial instanceof Node) {
 				node.attachChild(((Node) (spatial)).clone());
-			}
-			else {
+			} else {
 				new UnsupportedDataTypeException();
 			}
 		}
