+++ AL-Game/src/com/aionemu/gameserver/model/utils3d/Point3D.java
@@ -14,6 +14,7 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.aionemu.gameserver.model.utils3d;
 
 /**
@@ -38,9 +39,9 @@ public class Point3D {
 	}
 
 	public Point3D(float x, float y, float z) {
-		this.x = x;
-		this.y = y;
-		this.z = z;
+		this.x = (double) x;
+		this.y = (double) y;
+		this.z = (double) z;
 	}
 
 	public double distance(Point3D p) {
