+++ AL-Game/src/com/aionemu/gameserver/network/aion/serverpackets/SM_HOUSE_OWNER_INFO.java
@@ -14,11 +14,8 @@
  *  along with Aion-Lightning.
  *  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.aionemu.gameserver.network.aion.serverpackets;
-
-import java.sql.Timestamp;
-
-import org.joda.time.DateTime;
 
 import com.aionemu.gameserver.model.gameobjects.player.Player;
 import com.aionemu.gameserver.model.gameobjects.player.PlayerHouseOwnerFlags;
@@ -29,6 +26,10 @@ import com.aionemu.gameserver.model.town.Town;
 import com.aionemu.gameserver.network.aion.AionConnection;
 import com.aionemu.gameserver.network.aion.AionServerPacket;
 import com.aionemu.gameserver.services.TownService;
+
+import org.joda.time.DateTime;
+
+import java.sql.Timestamp;
 
 /**
  * @author Rolandas
@@ -48,8 +49,7 @@ public class SM_HOUSE_OWNER_INFO extends AionServerPacket {
 		if (activeHouse == null) {
 			writeD(0);
 			writeD(player.isBuildingInState(PlayerHouseOwnerFlags.BUY_STUDIO_ALLOWED) ? 355000 : 0); // studio building id
-		}
-		else {
+		} else {
 			writeD(activeHouse.getAddress().getId());
 			writeD(activeHouse.getBuilding().getId());
 		}
@@ -63,8 +63,7 @@ public class SM_HOUSE_OWNER_INFO extends AionServerPacket {
 		// Maintenance bill weeks left ?, if 0 maintenance date is in red
 		if (activeHouse == null || !activeHouse.isFeePaid() || activeHouse.getHouseType() == HouseType.STUDIO) {
 			writeC(0);
-		}
-		else {
+		} else {
 			Timestamp nextPay = activeHouse.getNextPay();
 			float diff;
 			if (nextPay == null) {
@@ -72,16 +71,14 @@ public class SM_HOUSE_OWNER_INFO extends AionServerPacket {
 				// all just obtained houses have fee paid true and time is null;
 				// means they should pay next week
 				diff = MaintenanceTask.getInstance().getPeriod();
-			}
-			else {
+			} else {
 				long paytime = activeHouse.getNextPay().getTime();
 				diff = paytime - ((long) MaintenanceTask.getInstance().getRunTime() * 1000);
 			}
 			if (diff < 0) {
 				writeC(0);
-			}
-			else {
-				int weeks = (Math.round(diff / MaintenanceTask.getInstance().getPeriod()));
+			} else {
+				int weeks = (int) (Math.round(diff / MaintenanceTask.getInstance().getPeriod()));
 				if (DateTime.now().getDayOfWeek() != 7) // Hack for auction Day, client counts sunday to new week
 				{
 					weeks++;
